requirements.txt
django
gunicorn
psycopg2-binary
whitenoise
pandas
dj_database_url
pytest
dotenv
djangorestframework
djangorestframework-simplejwt
polars